tests/wpt/web-platform-tests/css/work-in-progress/gadams/cssom/cssstyledeclaration-instance-named-properties-in-css20-only.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD//XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml">
  <head>
    <title>CSSOM - CSSStyleDeclaration - Inline Style - Named Properties - CSS2.0 Only - Non-Aural</title>
    <link rel="author" title="Glenn Adams" href="mailto:glenn@skynav.com"/>
    <link rel="help" href="http://www.w3.org/TR/cssom-1/#the-cssstyledeclaration-interface"/>
    <meta name="flags" content="dom"/>
    <script src="/resources/testharness.js"/>
    <script src="/resources/testharnessreport.js"/>
  </head>
  <body>
    <div id="log"></div>
    <div id="box"></div>
    <script><![CDATA[

      var elt = document.getElementById('box');

      test(function(){

        assert_true('style' in elt);

      }, '\'style\' in elt');

      var style = elt.style;

      test(function(){

        assert_true('fontSizeAdjust' in style);

      }, '\'fontSizeAdjust\' in style');

      test(function(){

        assert_true('fontStretch' in style);

      }, '\'fontStretch\' in style');

      test(function(){

        assert_true('markerOffset' in style);

      }, '\'markerOffset\' in style');

      test(function(){

        assert_true('marks' in style);

      }, '\'marks\' in style');

      test(function(){

        assert_true('page' in style);

      }, '\'page\' in style');

      test(function(){

        assert_true('size' in style);

      }, '\'size\' in style');

      test(function(){

        assert_true('textShadow' in style);

      }, '\'textShadow\' in style');

    ]]>
    </script>
  </body>
</html>
